+++ NEWS.md
@@ -1,8 +1,9 @@
 # whirl dev
 
-* Add `write_biocompute()` to create [BioCompute Objects](https://www.biocomputeobject.org/) containing the logs in a standardised JSON format.
-* Calling `run()` with `track_files = TRUE` now checks if strace can be attached to the process
-
+* Add `write_biocompute()` to create [BioCompute Objects](https://www.biocomputeobject.org/) containing the logs in a standardized JSON format.
+* Calling `run()` with `track_files = TRUE` now checks if strace can be attached to the process.
+* Improved json logs and similar returned output from running a script (`result` column in return from `run()`).
+* Simplified approved packages check. Now the user supplies a character vector of packages and versions specified as `{package}@{version}`.
 
 # whirl 0.2.0